test: 게시글 조회 로직 추가로 인한 test code 작성
